+++ .settings/org.eclipse.core.resources.prefs
@@ -2,10 +2,6 @@ eclipse.preferences.version=1
 encoding//Debug/makefile=UTF-8
 encoding//Debug/objects.mk=UTF-8
 encoding//Debug/sources.mk=UTF-8
-encoding//Debug/src/dsp/subdir_rules.mk=UTF-8
-encoding//Debug/src/dsp/subdir_vars.mk=UTF-8
-encoding//Debug/src/platform/subdir_rules.mk=UTF-8
-encoding//Debug/src/platform/subdir_vars.mk=UTF-8
 encoding//Debug/src/subdir_rules.mk=UTF-8
 encoding//Debug/src/subdir_vars.mk=UTF-8
 encoding//Debug/subdir_rules.mk=UTF-8
