+++ Debug/ccsObjs.opt
@@ -1,1 +1,1 @@
--l"./configPkg/linker.cmd" "./src/dsp/gen_twiddle_fft32x32.obj" "./src/dsp/gen_twiddle_fft16x32.obj" "./src/dsp/gen_twiddle_fft16x16.obj" "./src/dsp/fftTest.obj" "./src/main.obj" "./src/Utils.obj" "./src/Thread_NetworkInit.obj" "./src/Thread_ADS1298Init.obj" "./src/Services.obj" "./src/SPI_Intc.obj" "./src/KeyStone_common.obj" "./src/KeyStone_UART_Init_drv.obj" "./src/KeyStone_SPI_Init_drv.obj" "./src/KeyStone_DDR_Init.obj" "./src/GPIO_drv.obj" "./src/ADS1298_drv.obj" "./src/platform/resourcemgr.obj" "./src/platform/platform_osal.obj" -l"libc.a" -l"D:\ti\pdk_C6678_1_1_2_5\packages\ti\platform\evmc6678l\platform_lib\lib\release\ti.platform.evm6678l.ae66" -l"D:\ti\bios_6_33_06_50\packages\ti\sysbios\family\c66\tci66xx\lib\debug\ti.sysbios.family.c66.tci66xx.ae66" -l"D:\ti\pdk_C6678_1_1_2_5\packages\ti\csl\lib\ti.csl.ae66" -l"D:\ti\pdk_C6678_1_1_2_5\packages\ti\csl\lib\ti.csl.intc.ae66" 
+-l"./configPkg/linker.cmd" "./src/resourcemgr.obj" "./src/platform_osal.obj" "./src/main.obj" "./src/Utils.obj" "./src/Thread_NetworkInit.obj" "./src/Thread_ADS1298Init.obj" "./src/Services.obj" "./src/SPI_Intc.obj" "./src/KeyStone_common.obj" "./src/KeyStone_UART_Init_drv.obj" "./src/KeyStone_SPI_Init_drv.obj" "./src/KeyStone_DDR_Init.obj" "./src/GPIO_drv.obj" "./src/ADS1298_drv.obj" -l"libc.a" -l"D:\ti\pdk_C6678_1_1_2_5\packages\ti\platform\evmc6678l\platform_lib\lib\release\ti.platform.evm6678l.ae66" -l"D:\ti\bios_6_33_06_50\packages\ti\sysbios\family\c66\tci66xx\lib\debug\ti.sysbios.family.c66.tci66xx.ae66" -l"D:\ti\pdk_C6678_1_1_2_5\packages\ti\csl\lib\ti.csl.ae66" -l"D:\ti\pdk_C6678_1_1_2_5\packages\ti\csl\lib\ti.csl.intc.ae66"
