+++ Debug/configPkg/compiler.opt
@@ -1,1 +1,1 @@
--mv6600 --abi=eabi -I"D:/ti/pdk_C6678_1_1_2_5/packages" -I"D:/ti/ndk_2_21_01_38/packages" -I"D:/ti/bios_6_33_06_50/packages" -I"D:/ti/mcsdk_2_01_02_05/demos" -I"D:/ti/ccsv5/ccs_base" -I"D:/ti/dsplib_c66x_3_1_0_0/packages" -I"D:/ti/xdctools_3_23_04_60/packages" -I"D:/CCS_PRJ/Pacemaker_DET/Debug/configPkg/.." -IC:/PROGRA~1/TEXASI~2/C6000C~1.6/include -Dxdc_target_types__="ti/targets/elf/std.h" -Dxdc_target_name__=C66 -Dxdc_cfg__header__="D:/CCS_PRJ/Pacemaker_DET/Debug/configPkg/package/cfg/Pacemaker_DET_pe66.h"
+-mv6600 --abi=eabi -I"D:/ti/pdk_C6678_1_1_2_5/packages" -I"D:/ti/ndk_2_21_01_38/packages" -I"D:/ti/bios_6_33_06_50/packages" -I"D:/ti/mcsdk_2_01_02_05/demos" -I"D:/ti/ccsv5/ccs_base" -I"D:/ti/xdctools_3_23_04_60/packages" -I"D:/CCS_PRJ/Pacemaker_DET/Debug/configPkg/.." -IC:/PROGRA~1/TEXASI~2/C6000C~1.6/include -Dxdc_target_types__="ti/targets/elf/std.h" -Dxdc_target_name__=C66 -Dxdc_cfg__header__="D:/CCS_PRJ/Pacemaker_DET/Debug/configPkg/package/cfg/Pacemaker_DET_pe66.h"
