+++ Debug/sources.mk
@@ -108,7 +108,5 @@ S64_DEPS__QUOTED :=
 # Every subdirectory with source files must be described here
 SUBDIRS := \
 . \
-src/platform \
 src \
-src/dsp \
