+++ Debug/src/ccsSrcs.opt
@@ -1,1 +1,1 @@
-"../src/ADS1298_drv.c" "../src/GPIO_drv.c" "../src/KeyStone_DDR_Init.c" "../src/KeyStone_SPI_Init_drv.c" "../src/KeyStone_UART_Init_drv.c" "../src/KeyStone_common.c" "../src/SPI_Intc.c" "../src/Services.c" "../src/Thread_ADS1298Init.c" "../src/Thread_NetworkInit.c" "../src/Utils.c" "../src/main.c" 
+"../src/ADS1298_drv.c" "../src/GPIO_drv.c" "../src/KeyStone_DDR_Init.c" "../src/KeyStone_SPI_Init_drv.c" "../src/KeyStone_UART_Init_drv.c" "../src/KeyStone_common.c" "../src/SPI_Intc.c" "../src/Services.c" "../src/Thread_ADS1298Init.c" "../src/Thread_NetworkInit.c" "../src/Utils.c" "../src/main.c" "../src/platform_osal.c" "../src/resourcemgr.c"
