--- Debug/src/dsp/ccsSrcs.opt
@@ -1,1 +1,0 @@
-"../src/dsp/fftTest.c" "../src/dsp/gen_twiddle_fft16x16.c" "../src/dsp/gen_twiddle_fft16x32.c" "../src/dsp/gen_twiddle_fft32x32.c"
