--- Debug/src/dsp/subdir_vars.mk
@@ -1,42 +1,0 @@
-################################################################################
-# Automatically-generated file. Do not edit!
-################################################################################
-
-# Add inputs and outputs from these tool invocations to the build variables 
-C_SRCS += \
-../src/dsp/fftTest.c \
-../src/dsp/gen_twiddle_fft16x16.c \
-../src/dsp/gen_twiddle_fft16x32.c \
-../src/dsp/gen_twiddle_fft32x32.c 
-
-OBJS += \
-./src/dsp/fftTest.obj \
-./src/dsp/gen_twiddle_fft16x16.obj \
-./src/dsp/gen_twiddle_fft16x32.obj \
-./src/dsp/gen_twiddle_fft32x32.obj 
-
-C_DEPS += \
-./src/dsp/fftTest.pp \
-./src/dsp/gen_twiddle_fft16x16.pp \
-./src/dsp/gen_twiddle_fft16x32.pp \
-./src/dsp/gen_twiddle_fft32x32.pp 
-
-C_DEPS__QUOTED += \
-"src\dsp\fftTest.pp" \
-"src\dsp\gen_twiddle_fft16x16.pp" \
-"src\dsp\gen_twiddle_fft16x32.pp" \
-"src\dsp\gen_twiddle_fft32x32.pp" 
-
-OBJS__QUOTED += \
-"src\dsp\fftTest.obj" \
-"src\dsp\gen_twiddle_fft16x16.obj" \
-"src\dsp\gen_twiddle_fft16x32.obj" \
-"src\dsp\gen_twiddle_fft32x32.obj" 
-
-C_SRCS__QUOTED += \
-"../src/dsp/fftTest.c" \
-"../src/dsp/gen_twiddle_fft16x16.c" \
-"../src/dsp/gen_twiddle_fft16x32.c" \
-"../src/dsp/gen_twiddle_fft32x32.c" 
-
-
