+++ Debug/src/main.pp
@@ -104,8 +104,6 @@ src/main.obj: D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr_device.h
 src/main.obj: D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr.h
 src/main.obj: D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/tistdtypes.h
 src/main.obj: D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr_uart.h
-src/main.obj: D:/CCS_PRJ/Pacemaker_DET/include/KeyStone_DDR_Init.h
-src/main.obj: D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr_emif4f.h
 src/main.obj: D:/CCS_PRJ/Pacemaker_DET/include/KeyStone_common.h
 src/main.obj: C:/Program\ Files/Texas\ Instruments/C6000\ Code\ Generation\ Tools\ 7.4.6/include/c6x.h
 src/main.obj: C:/Program\ Files/Texas\ Instruments/C6000\ Code\ Generation\ Tools\ 7.4.6/include/vect.h
@@ -136,6 +134,7 @@ src/main.obj: D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr_msmc.h
 src/main.obj: D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/csl_idma.h
 src/main.obj: D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr_mpu.h
 src/main.obj: D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr_gpio.h
+src/main.obj: D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr_emif4f.h
 src/main.obj: C:/Program\ Files/Texas\ Instruments/C6000\ Code\ Generation\ Tools\ 7.4.6/include/stdio.h
 src/main.obj: D:/CCS_PRJ/Pacemaker_DET/include/Global.h
 
@@ -243,8 +242,6 @@ D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr_device.h:
 D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr.h: 
 D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/tistdtypes.h: 
 D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr_uart.h: 
-D:/CCS_PRJ/Pacemaker_DET/include/KeyStone_DDR_Init.h: 
-D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr_emif4f.h: 
 D:/CCS_PRJ/Pacemaker_DET/include/KeyStone_common.h: 
 C:/Program\ Files/Texas\ Instruments/C6000\ Code\ Generation\ Tools\ 7.4.6/include/c6x.h: 
 C:/Program\ Files/Texas\ Instruments/C6000\ Code\ Generation\ Tools\ 7.4.6/include/vect.h: 
@@ -275,5 +272,6 @@ D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr_msmc.h:
 D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/csl_idma.h: 
 D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr_mpu.h: 
 D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr_gpio.h: 
+D:/ti/pdk_C6678_1_1_2_5/packages/ti/csl/cslr_emif4f.h: 
 C:/Program\ Files/Texas\ Instruments/C6000\ Code\ Generation\ Tools\ 7.4.6/include/stdio.h: 
 D:/CCS_PRJ/Pacemaker_DET/include/Global.h:
