--- Debug/src/platform/ccsSrcs.opt
@@ -1,1 +1,0 @@
-"../src/platform/platform_osal.c" "../src/platform/resourcemgr.c"
