--- Debug/src/platform/subdir_vars.mk
@@ -1,30 +1,0 @@
-################################################################################
-# Automatically-generated file. Do not edit!
-################################################################################
-
-# Add inputs and outputs from these tool invocations to the build variables 
-C_SRCS += \
-../src/platform/platform_osal.c \
-../src/platform/resourcemgr.c 
-
-OBJS += \
-./src/platform/platform_osal.obj \
-./src/platform/resourcemgr.obj 
-
-C_DEPS += \
-./src/platform/platform_osal.pp \
-./src/platform/resourcemgr.pp 
-
-C_DEPS__QUOTED += \
-"src\platform\platform_osal.pp" \
-"src\platform\resourcemgr.pp" 
-
-OBJS__QUOTED += \
-"src\platform\platform_osal.obj" \
-"src\platform\resourcemgr.obj" 
-
-C_SRCS__QUOTED += \
-"../src/platform/platform_osal.c" \
-"../src/platform/resourcemgr.c" 
-
-
