--- include/dsp.h
@@ -1,13 +1,0 @@
-/*
- * dsp.h
- *
- *  Created on: 2016-1-14
- *      Author: Wayne
- */
-
-#ifndef DSP_H_
-#define DSP_H_
-
-extern void signal_process();
-
-#endif /* DSP_H_ */
